+++ src/harlequin/components/help_screen.py
@@ -7,10 +7,16 @@ from textual.screen import ModalScreen
 from textual.widgets import Markdown, Static
 
 
+class VerticalSuppressClicks(Vertical):
+    def on_click(self, message: events.Click) -> None:
+        message.stop()
+
+
 class HelpScreen(ModalScreen):
     header_text = """
         Harlequin supports a number of key bindings ("shortcuts"). The list below
         includes all of the bindings supported by components of the app.
+        You can also view this list at https://harlequin.sh/docs/bindings
     """.split()
 
     def compose(self) -> ComposeResult:
@@ -18,7 +24,7 @@ class HelpScreen(ModalScreen):
         with open(markdown_path, "r") as f:
             markdown = f.read()
 
-        with Vertical(id="help_outer"):
+        with VerticalSuppressClicks(id="help_outer"):
             yield Static(" ".join(self.header_text), id="help_header")
             with VerticalScroll(id="help_inner"):
                 yield Markdown(markdown=markdown)
@@ -47,3 +53,6 @@ class HelpScreen(ModalScreen):
             self.body.scroll_page_down()
         else:
             self.app.pop_screen()
+
+    def on_click(self) -> None:
+        self.app.pop_screen()