+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Features
 
 -   Double-click or press <kbd>ctrl+enter</kbd> on an item in the data catalog to insert the name in the query editor ([#194](https://github.com/tconbeer/harlequin/issues/194)).
+-   Harlequin now shows notifications when exporting data or executing DDL/DML.
 
 ### Bug Fixes and Minor Updates
 
@@ -14,6 +15,7 @@ All notable changes to this project will be documented in this file.
 -   Bumped TextArea; cursor now better maintains x-position and [other minor fixes](https://github.com/tconbeer/textual-textarea/releases/tag/v0.5.4).
 -   The query editor's cursor no longer blinks when a modal appears above it ([#196](https://github.com/tconbeer/harlequin/issues/196)).
 -   Harlequin now shows the results of successful queries in the Results Viewer if multiple queries are executed and one or more contain errors.
+-   Error and Help modals can now be dismissed with a click outside the modal ([#218](https://github.com/tconbeer/harlequin/issues/218)).
 
 ## [0.0.28] - 2023-09-07